+++ src/effects/RandomColors.cpp
@@ -14,7 +14,7 @@ bool RandomColors::loop(Hexes *hexes)
 		return false;
 	}
 
-	for (int i = 0; i < hexes->num_hexes; i++)
+	for (size_t i = 0; i < hexes->num_hexes; i++)
 	{
 		Hex *hex = hexes->get_by_index(i);
